{"schema":"decodex/commit/1","summary":"Move macOS key focus shell helpers","authority":"manual"}
